+++ yukon-client/yukon-web/src/com/cannontech/web/api/der/edge/service/DerEdgeResponseService.java
@@ -2,5 +2,5 @@ package com.cannontech.web.api.der.edge.service;
 
 public interface DerEdgeResponseService {
 
-    void addCacheEntry(Short e2eID, String guid);
+    void addCacheEntry(Integer e2eID, String guid);
 }
